+++ src/Events.php
@@ -16,7 +16,7 @@ namespace FOS\HttpCache;
  */
 final class Events
 {
-    const PROXY_UNREACHABLE_ERROR = 'fos_http_cache.error.proxy_unreachable';
+    public const PROXY_UNREACHABLE_ERROR = 'fos_http_cache.error.proxy_unreachable';
 
-    const PROXY_RESPONSE_ERROR = 'fos_http_cache.error.response';
+    public const PROXY_RESPONSE_ERROR = 'fos_http_cache.error.response';
 }