+++ src/ProxyClient/Nginx.php
@@ -25,11 +25,11 @@ use FOS\HttpCache\ProxyClient\Invalidation\RefreshCapable;
  */
 class Nginx extends HttpProxyClient implements PurgeCapable, RefreshCapable
 {
-    const HTTP_METHOD_PURGE = 'PURGE';
+    public const HTTP_METHOD_PURGE = 'PURGE';
 
-    const HTTP_METHOD_REFRESH = 'GET';
+    public const HTTP_METHOD_REFRESH = 'GET';
 
-    const HTTP_HEADER_REFRESH = 'X-Refresh';
+    public const HTTP_HEADER_REFRESH = 'X-Refresh';
 
     /**
      * {@inheritdoc}